+++ dart/flutter_monet_theme/lib/flutter_monet_theme.dart
@@ -1,6 +1,7 @@
 library flutter_monet_theme;
 
 export 'src/baseline.dart';
+export 'src/custom_color.dart';
 export 'src/device.dart';
 export 'src/elevation.dart';
 export 'src/generate.dart';