+++ dart/flutter_monet_theme/lib/src/generate.dart
@@ -26,3 +26,8 @@ MonetTheme generateTheme(
         errorSeed: errorSeed?.value,
       ),
     );
+
+CustomColorTheme generateCustomColorThemeFrom(ColorTonalPalette tonalPalette) =>
+    CustomColorTheme.fromRaw(
+      generateRawCustomColorThemeFrom(tonalPalette.raw),
+    );